requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
gunicorn==23.0.0
ImageHash==4.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.3
packaging==25.0
piexif==1.1.3
pillow==11.3.0
PyWavelets==1.9.0
scipy==1.16.2
Werkzeug==3.1.3
